+++ common/mount/__init__.py
@@ -453,16 +453,16 @@ class MountManager:
             return
 
         if not self._lock_mountpoint.exists():
-            # DEBUG
-            # pylint: disable-next=import-outside-toplevel
-            import traceback  # noqa: PLC0415
-            traceback.print_stack(limit=5)
+            # # DEBUG
+            # # pylint: disable-next=import-outside-toplevel
+            # import traceback  # no-qa: PLC0415
+            # traceback.print_stack(limit=5)
 
-            logger.warning(
-                'Mount point lock - Unexpected state. '
-                f'{self._lock_mountpoint} '
-                'does not exist.', self
-            )
+            # logger.warning(
+            #     'Mount point lock - Unexpected state. '
+            #     f'{self._lock_mountpoint} '
+            #     'does not exist.', self
+            # )
             return
 
         logger.debug(